requirements.txt
appnope==0.1.4
asttokens==3.0.0
cloudpickle==3.1.0
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nibabel==5.3.2
numpy==2.2.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pyzmq==26.2.0
six==1.17.0
spyder-kernels==3.0.2
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
wcwidth==0.2.13
wurlitzer==3.1.1
